Refactor : 더미데이터 삭제, 프롭스추가
